+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - [BREAKING] Added `BlockNumber` struct (#1043, #1080, #1082).
 - [BREAKING] Removed `GENESIS_BLOCK` public constant (#1088).
 - Add CI check for unused dependencies (#1075).
+- Added storage placeholder types and support for templated map (#1074).
 
 ## 0.6.2 (2024-11-20)